Migrate ?link to slash command
