+++ src/events/messageCreate/commands/index.ts
@@ -5,7 +5,6 @@ import ping from "./dev/ping";
 import dump from "./dev/dump";
 import introduce from "./help/introduce";
 import warn from "./moderation/warn";
-import link from "./help/link";
 import rankup from "./fun/rankup";
 
 const commands: { [k: string]: Command } = {
@@ -13,7 +12,6 @@ const commands: { [k: string]: Command } = {
   dump,
   introduce,
   warn,
-  link,
   rankup,
 };
 export default commands;
